python/requirements.txt
kafka-python
flask
lxml
xmltodict
xmldiff
#psycopg2
psycopg2-binary
pytest
emoji
streamlit